requirements.txt
PyYAML=3.13
git+git://github.com/Julian/jsonschema.git@v3.0.0a2
